Expose corc_alloc_bad_cb() and check for verbose.
corc.alloc checks at compile-time for the existence of the CORC_VERBOSE
definition. If this exists, the default callback for corc_alloc_bad_cb
will print a short statement to STDERR informing that memory allocation
failed and that the program is now exiting. This is only enabled upon
existence of the CORC_VERBOSE definition (In the C preprocessor), and
if this definition does not exist, stdio.h will not be included in and
bad_callback() will not print a message, instead opting to silently
exit.

corc.alloc sets the allocation failure callback to bad_callback() by
default. This is useful as it presents a [sane] action for when malloc
issues come up. The program using this library is capable of changing
the callback via passing a void-typed function to corc_alloc_bad_cb().

We have chosen to keep corc_free() in anticipation of future API
changes. At present, corc.alloc is a very basic wrapper around malloc,
but in the future, there are plans for it to become a bit smarter of a
function. In that case, corc_free() will be needed, therefore we should
start using it now instead of later.